correciones solucionadas (#644)

+++ TP/62263 - Orellana, José Augusto/tp6/servidor/Program.cs
@@ -19,9 +19,13 @@ builder.Services.AddDbContext<TiendaContext>(options =>
 var app = builder.Build();
 
 // Configurar el pipeline de solicitudes HTTP
-if (app.Environment.IsDevelopment()) {
+if (app.Environment.IsDevelopment())
+{
     app.UseDeveloperExceptionPage();
 }
+
+// Para habilitar archivos estáticos
+app.UseStaticFiles();
 
 // Usar CORS con la política definida
 app.UseCors("AllowClientApp");
@@ -43,16 +47,16 @@ using (var scope = app.Services.CreateScope())
     {
         var productosIniciales = new List<Producto>
         {
-            new Producto { Nombre = "Celular Galaxy A54", Descripcion = "Celular Samsung 128GB", Precio = 190000, Stock = 10, ImagenUrl = "https://imgs.search.brave.com/LaLRpDspuUpj9USNOkcKXx4Jgm99wY3O9zS3z8AIMCs/rs:fit:860:0:0:0/g:ce/aHR0cHM6Ly9jZG4u/bGVzbnVtZXJpcXVl/cy5jb20vb3B0aW0v/cHJvZHVjdC83MS83/MTk3MS85YjU3YWQw/OC1nYWxheHktYTU0/X180MDBfNDAwLmpw/Zw" },
-            new Producto { Nombre = "Celular iPhone 13", Descripcion = "Apple iPhone 13 256GB", Precio = 420000, Stock = 5, ImagenUrl = "https://imgs.search.brave.com/aoxrpNi0vTljtdxmNM4WmsYsq6vftQtByDl0chNHr7g/rs:fit:860:0:0:0/g:ce/aHR0cHM6Ly90aWVu/ZGEucGVyc29uYWwu/Y29tLmFyL2ltYWdl/cy83MjAvd2VicC9p/X1Bob25lXzEzX1By/b18yNTZfR0JfR3Jh/cGhpdGVfNzNmM2Uw/NmQxYy5wbmc" },
-            new Producto { Nombre = "Cargador USB-C", Descripcion = "Cargador rápido 20W", Precio = 7000, Stock = 30, ImagenUrl = "https://imgs.search.brave.com/hmXX10qrpBgPf9pdr2tS10AkXFlA7-PhUk7H5CB1hvs/rs:fit:860:0:0:0/g:ce/aHR0cHM6Ly9odHRw/Mi5tbHN0YXRpYy5j/b20vRF9RX05QXzJY/Xzg0Njg3Ni1NTFU3/NTg1NTU4ODU1N18w/NDIwMjQtVi53ZWJw" },
-            new Producto { Nombre = "Auriculares Bluetooth", Descripcion = "Auriculares inalámbricos", Precio = 15000, Stock = 25, ImagenUrl = "https://imgs.search.brave.com/NrkV8_8ym_q582Nb4t9CjbSkMRFulRNzV47T7zIBMLQ/rs:fit:860:0:0:0/g:ce/aHR0cHM6Ly90aWVu/ZGEucGVyc29uYWwu/Y29tLmFyL2ltYWdl/cy83MjAvd2VicC9B/dXJpY3VsYXJlc19C/bHVldG9vdGhfRm94/Ym94X09wZW5haXJf/Q2xpcF9jMTdlY2Ux/ZWJhLnBuZw" },
-            new Producto { Nombre = "Smartwatch", Descripcion = "Reloj inteligente con GPS", Precio = 25000, Stock = 12, ImagenUrl = "https://imgs.search.brave.com/ICfA_LH4vnplJ6q30uPq2RSeFR6ro0m4XodiDM_baSw/rs:fit:860:0:0:0/g:ce/aHR0cHM6Ly93d3cu/dGl0YW4uY28uaW4v/ZHcvaW1hZ2UvdjIv/QktERF9QUkQvb24v/ZGVtYW5kd2FyZS5z/dGF0aWMvLS9TaXRl/cy10aXRhbi1tYXN0/ZXItY2F0YWxvZy9k/ZWZhdWx0L2R3YTZj/OGE5MWUvaW1hZ2Vz/L0Zhc3RyYWNrL0Nh/dGFsb2cvMzgxMjNR/TTAxXzQuanBnP3N3/PTM2MCZzaD0zNjA" },
-            new Producto { Nombre = "Notebook Lenovo", Descripcion = "Laptop 15.6'' 8GB RAM", Precio = 320000, Stock = 7, ImagenUrl = "https://imgs.search.brave.com/442IklAxpbCqzYRhJ_mYwoHHJiPfp1Ou5abe1kJTHPE/rs:fit:860:0:0:0/g:ce/aHR0cHM6Ly9odHRw/Mi5tbHN0YXRpYy5j/b20vRF9RX05QXzJY/Xzg2MTgxMy1NTFU3/NzEyNjkxNTYxOF8w/NjIwMjQtRS53ZWJw" },
-            new Producto { Nombre = "Gaseosa Coca-Cola 2L", Descripcion = "Botella 2 litros", Precio = 1000, Stock = 50, ImagenUrl = "https://imgs.search.brave.com/cMOCtFIHhrcTIWd-pQMVnZihbOLYs-LQ2qa10p2-q8I/rs:fit:860:0:0:0/g:ce/aHR0cHM6Ly9hbmRp/bmFjb2NhY29sYWFy/LnZ0ZXhhc3NldHMu/Y29tL2FycXVpdm9z/L2lkcy8xNTYyMjMt/ODAwLTgwMD92PTYz/ODU3MDE1NTkyNDM3/MDAwMCZ3aWR0aD04/MDAmaGVpZ2h0PTgw/MCZhc3BlY3Q9dHJ1/ZQ" },
-            new Producto { Nombre = "Gaseosa Pepsi 2L", Descripcion = "Pepsi 2 litros", Precio = 950, Stock = 45, ImagenUrl = "https://imgs.search.brave.com/ZAjNsfIYrtqJy0RJdPHH8s4Omy0mSAXevb8c6LRD_Gc/rs:fit:860:0:0:0/g:ce/aHR0cHM6Ly9tLm1l/ZGlhLWFtYXpvbi5j/b20vaW1hZ2VzL0kv/NTFZRTlqcFNjUkwu/anBn" },
-            new Producto { Nombre = "Chocolate Milka", Descripcion = "Chocolate con leche 100g", Precio = 800, Stock = 100, ImagenUrl = "https://elnenearg.vtexassets.com/arquivos/ids/164041-800-auto?v=638023197840300000&width=800&height=auto&aspect=true" },
-            new Producto { Nombre = "Papas Lays", Descripcion = "Bolsa de papas fritas 200g", Precio = 1200, Stock = 60, ImagenUrl = "https://imgs.search.brave.com/_2c-50Qz6ktaToRAjHUjhqBUMMh2iPiq7jyklAmDW84/rs:fit:860:0:0:0/g:ce/aHR0cHM6Ly90dXN1/cGVyLmNvbS5hci9p/bWFnZS9jYWNoZS9j/YXRhbG9nL1AyMDIz/L0RpY2llbWJyZSUy/MDE2L1BhcGFzLUZy/aXRhcy1DbGFzaWNh/cy1MYXlzLTg1LUdy/LTEtMTY5NzAlMjAo/MSklMjAoMSktNjAw/eDYwMC5qcGc" },
+            new Producto { Nombre = "Celular Galaxy A54", Descripcion = "Celular Samsung 128GB", Precio = 190000, Stock = 10, ImagenUrl = "/images/galaxy.webp" },
+            new Producto { Nombre = "Celular iPhone 13", Descripcion = "Apple iPhone 13 256GB", Precio = 420000, Stock = 5, ImagenUrl = "/images/iPhone.webp" },
+            new Producto { Nombre = "Cargador USB-C", Descripcion = "Cargador rápido 20W", Precio = 7000, Stock = 30, ImagenUrl = "/images/cargador.webp" },
+            new Producto { Nombre = "Auriculares Bluetooth", Descripcion = "Auriculares inalámbricos", Precio = 15000, Stock = 25, ImagenUrl = "/images/auriculares.webp" },
+            new Producto { Nombre = "Smartwatch", Descripcion = "Reloj inteligente con GPS", Precio = 25000, Stock = 12, ImagenUrl = "/images/smartwatch.webp" },
+            new Producto { Nombre = "Notebook Lenovo", Descripcion = "Laptop 15.6'' 8GB RAM", Precio = 320000, Stock = 7, ImagenUrl = "/images/notebook.webp" },
+            new Producto { Nombre = "Gaseosa Coca-Cola 2L", Descripcion = "Botella 2 litros", Precio = 1000, Stock = 50, ImagenUrl = "/images/coca.webp" },
+            new Producto { Nombre = "Gaseosa Pepsi 2L", Descripcion = "Pepsi 2 litros", Precio = 950, Stock = 45, ImagenUrl = "/images/pepsi.webp" },
+            new Producto { Nombre = "Chocolate Milka", Descripcion = "Chocolate con leche 100g", Precio = 800, Stock = 100, ImagenUrl = "/images/milka.webp" },
+            new Producto { Nombre = "Papas Lays", Descripcion = "Bolsa de papas fritas 200g", Precio = 1200, Stock = 60, ImagenUrl = "/images/lays.webp" },
         };
 
         db.Productos.AddRange(productosIniciales);
@@ -127,17 +131,25 @@ app.MapDelete("/carritos/{carritoId:int}", async (int carritoId, TiendaContext d
 {
     var compra = await db.Compras
         .Include(c => c.Items)
+        .ThenInclude(i => i.Producto)
         .FirstOrDefaultAsync(c => c.Id == carritoId);
 
     if (compra == null) return Results.NotFound("Carrito no encontrado");
 
+    // Se devuelve la cantidad al stock de cada producto
+    foreach (var item in compra.Items)
+    {
+        item.Producto.Stock += item.Cantidad;
+    }
+
+    // Se eliminan los ítems del carrito
     db.ItemsCompra.RemoveRange(compra.Items);
     await db.SaveChangesAsync();
 
     return Results.Ok("Carrito vaciado exitosamente");
 });
 
-// Endpoint para confirmar una carrito de compras
+// Endpoint para confirmar un carrito de compras
 app.MapPut("/carritos/{carritoId}/confirmar", async (int carritoId, ClienteDTO cliente, TiendaContext db) =>
 {
     // Validar campos del cliente
@@ -167,16 +179,10 @@ app.MapPut("/carritos/{carritoId}/confirmar", async (int carritoId, ClienteDTO c
         return Results.BadRequest("Este carrito ya fue confirmado anteriormente");
     }
 
-    foreach (var item in compra.Items)
-        {
-            if (item.Cantidad > item.Producto.Stock) return Results.BadRequest($"No hay stock suficiente para el producto {item.Producto.Nombre}");
-        }
-
-    // Actualizar stock de productos y calcular total
+    // Se calcula el total sin modificar el stock
     decimal total = 0;
     foreach (var item in compra.Items)
     {
-        item.Producto.Stock -= item.Cantidad;
         total += item.Cantidad * item.PrecioUnitario;
     }
 
@@ -205,6 +211,7 @@ app.MapPut("/carritos/{carritoId:int}/{productoId:int}", async (int carritoId, i
 
     var carrito = await db.Compras
         .Include(c => c.Items)
+        .ThenInclude(i => i.Producto)
         .FirstOrDefaultAsync(c => c.Id == carritoId);
 
     if (carrito == null) return Results.NotFound("Carrito no encontrado");
@@ -214,8 +221,10 @@ app.MapPut("/carritos/{carritoId:int}/{productoId:int}", async (int carritoId, i
 
     var item = carrito.Items.FirstOrDefault(i => i.ProductoId == productoId);
 
-    int cantidadTotal = (item?.Cantidad ?? 0) + data.Cantidad;
-    if (cantidadTotal > producto.Stock) return Results.BadRequest("Stock insuficiente");
+    if (data.Cantidad > producto.Stock) return Results.BadRequest("Stock insuficiente");
+
+    // Se descuenta del stock
+    producto.Stock -= data.Cantidad;
 
     if (item != null)
     {
@@ -247,6 +256,13 @@ app.MapDelete("/carritos/{carritoId:int}/{productoId:int}", async (int carritoId
 
     var item = carrito.Items.FirstOrDefault(i => i.ProductoId == productoId);
     if (item == null) return Results.NotFound("Producto no encontrado en el carrito.");
+
+    // Se carga el producto para actualizar stock
+    var producto = await db.Productos.FindAsync(productoId);
+    if (producto == null) return Results.NotFound("Producto no encontrado.");
+
+    // Se incrementa el stock
+    producto.Stock += 1;
 
     if (item.Cantidad > 1) item.Cantidad--;
     else db.ItemsCompra.Remove(item);
